app/src/main/kotlin/skoda_backend/controllers/UserController.kt
package com.example.controllers

import io.ktor.server.application.*
import io.ktor.server.routing.*
import io.ktor.server.response.*

fun Application.userRoutes() {
    routing {
        route("/users") {
            get {
                // Example logic to get users
                call.respondText("Fetching user list...")
            }
            post {
                // Example logic to create a user
                call.respondText("User created successfully!")
            }
            get("/{id}") {
                // Example logic to fetch a user by ID
                val userId = call.parameters["id"]
                call.respondText("Fetching user with ID: $userId")
            }
            put("/{id}") {
                // Example logic to update a user by ID
                val userId = call.parameters["id"]
                call.respondText("Updating user with ID: $userId")
            }
            delete("/{id}") {
                // Example logic to delete a user by ID
                val userId = call.parameters["id"]
                call.respondText("Deleting user with ID: $userId")
            }
        }
    }
}
